+++ src/app/handlers/drag/interfaces.ts
@@ -1,4 +1,5 @@
 export interface ICuiElementDetector {
     detect(x: Number, y: Number): [number, Element];
     setElements(elements: Element[]): void;
+    setThreshold(value: number): void;
 }