+++ src/core/models/colors.ts
@@ -1,4 +1,5 @@
-import { getRangeValue, } from "../utils/functions";
+import { getRangeValue, is, isInRange, } from "../utils/functions";
+import { CuiColorError } from "./errors";
 
 export class CuiColor {
     #red: number;
@@ -7,6 +8,23 @@ export class CuiColor {
     #alpha: number;
     constructor(red: number, green: number, blue: number, alpha?: number) {
         this.set(red, green, blue, alpha);
+    }
+
+    static create(colorStr: string): CuiColor {
+        if (!is(colorStr)) {
+            throw new CuiColorError(`Given color value is not in a valid format`)
+        }
+        let trimmed = colorStr.trim();
+        let parser: ColorParser = null;
+        if (colorStr.startsWith("#")) {
+            parser = new HexColorParser(colorStr);
+        } else if (colorStr.startsWith("rgb")) {
+            parser = new RgbColorParser(colorStr);
+        } else {
+            throw new CuiColorError("Unknown color string format");
+        }
+
+        return parser.parse();
     }
 
     set(red: number, green: number, blue: number, alpha?: number) {
@@ -102,4 +120,79 @@ export interface CuiColorTheme {
 export interface CuiColorPair {
     light: string,
     dark: string
+}
+
+
+interface ColorParser {
+    trim(): void;
+    isValid(): boolean;
+    parse(): CuiColor;
+}
+
+export class HexColorParser implements ColorParser {
+    #colorStr: string;
+    constructor(colorStr: string) {
+        this.#colorStr = colorStr;
+    }
+
+    trim(): void {
+        this.#colorStr = this.#colorStr ? this.#colorStr.trim() : undefined;
+        return
+    }
+
+    isValid(): boolean {
+        return is(this.#colorStr) && this.#colorStr.startsWith("#")
+    }
+    parse(): CuiColor {
+        let red: number = 0;
+        let blue: number = 0;
+        let green: number = 0;
+        let alpha: number = 1;
+        let length = this.#colorStr.length;
+        if (this.#colorStr.length === 4) {
+            red = parseInt(this.#colorStr[1] + this.#colorStr[1], 16);
+            green = parseInt(this.#colorStr[2] + this.#colorStr[2], 16);
+            blue = parseInt(this.#colorStr[3] + this.#colorStr[3], 16);
+        } else {
+            red = parseInt(this.#colorStr[1] + this.#colorStr[2], 16);
+            green = parseInt(this.#colorStr[3] + this.#colorStr[4], 16);
+            blue = parseInt(this.#colorStr[5] + this.#colorStr[6], 16);
+
+            if (length > 7) {
+                let alphaHex = parseInt(this.#colorStr[7] + this.#colorStr[8], 16);
+                alpha = parseFloat((alphaHex / 255).toFixed(2));
+            }
+        }
+        return new CuiColor(red, green, blue, alpha);
+    }
+}
+
+export class RgbColorParser implements ColorParser {
+    #colorStr: string;
+
+    constructor(colorStr: string) {
+        this.#colorStr = colorStr;
+    }
+
+    trim(): void {
+        this.#colorStr = is(this.#colorStr) ? this.#colorStr.trim() : undefined;
+    }
+
+    isValid(): boolean {
+        return is(this.#colorStr) && this.#colorStr.startsWith("rgb");
+    }
+
+    parse(): CuiColor {
+        let len = this.#colorStr.length;
+        let str = this.#colorStr.startsWith("rgba") ? this.#colorStr.substring(5, len - 1) : this.#colorStr.substring(4, len - 1);
+        let split = str.split(",");
+        if (!isInRange(split.length, 3, 4)) {
+            return undefined;
+        }
+        let red: number = parseInt(split[0]);
+        let green: number = parseInt(split[1]);
+        let blue: number = parseInt(split[2]);
+        let alpha: number = split.length === 4 ? parseFloat(split[3]) : 1
+        return new CuiColor(red, green, blue, alpha);
+    }
 }