+++ src/index.ts
@@ -1,6 +1,6 @@
 import { CuiInit } from './app/initializer';
 
-export const CUI_LIGHT_VERSION = "0.1.32";
+export const CUI_LIGHT_VERSION = "0.1.33";
 
 declare global {
     interface Window {
@@ -12,6 +12,7 @@ export * from "./core/models/interfaces";
 export * from './core/models/events';
 export * from './core/models/setup';
 export * from './core/models/utils';
+export * from './core/models/colors';
 export * from './app/initializer'
 export { CuiInstance } from './app/instance';